requirements.txt
aiogram==3.13.1
aiohttp==3.10.9
APScheduler==3.10.4
tortoise-orm==0.21.6
aerich==0.7.2
python-dotenv==1.0.0
